implemented feedback

+++ src/crepe/api/Transform.h
@@ -16,9 +16,9 @@ public:
 	//! Translation (shift)
 	vec2 position = {0, 0};
 	//! Rotation, in degrees clockwise
-	double rotation = 0;
+	float rotation = 0;
 	//! Multiplication factor
-	double scale = 0;
+	float scale = 0;
 
 protected:
 	/**
